+++ lite/kernels/opencl/elementwise_add_image_compute.cc
@@ -78,10 +78,10 @@ void ElementwiseAddImageCompute::Run() {
       default_convertor.InitImageDimInfoWith(out->dims());  // w, h
   auto y_img_shape = default_convertor.InitImageDimInfoWith(y->dims());
 
-  auto* x_img = x->data<uint16_t, cl::Image2D>();
-  auto* y_img = y->data<uint16_t, cl::Image2D>();
-  auto* out_img = out->mutable_data<uint16_t, cl::Image2D>(out_img_shape[0],
-                                                           out_img_shape[1]);
+  auto* x_img = x->data<half_t, cl::Image2D>();
+  auto* y_img = y->data<half_t, cl::Image2D>();
+  auto* out_img = out->mutable_data<half_t, cl::Image2D>(out_img_shape[0],
+                                                         out_img_shape[1]);
 
   VLOG(4) << "x_img_shape[w,h]:" << x_img_width << " " << x_img_height;
   VLOG(4) << "y_img_shape[w,h]:" << y_img_shape[0] << " " << y_img_shape[1];
